pyproject.toml
[tool.poetry]
name = "ze-brands-test"
version = "1.0.0"
description = "ZeBrands catalog backend"
authors = ["Raquel <tu@email.com>"]
readme = "README.md"
packages = [{ include = "app", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
sqlalchemy = "^2.0.30"
ulid-py = "^1.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.10.1"
python-jose = { version = "^3.5.0", extras = ["cryptography"] }
bcrypt = "^3.2.0"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
pydantic = {extras = ["email"], version = "^2.11.7"}
boto3 = "^1.40.25"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^1.1.0"
httpx = "^0.24"
ruff = "^0.12"
pre-commit = "^3.4"
black = "^24.10"
yoyo-migrations = "^9.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-exec-plugin.commands]
lint = "poetry run ruff check . --fix --config ruff.toml"
test = "poetry run pytest"
migrate = "poetry run python scripts/run_yoyo_migrations.py"